Add a generic mechanism for reacting to ETH transfers
The Bridge now has the mapping isBridgeReceiver from address to bool to
indicate which contracts are whitelisted as bridge receivers.

These contracts need to implement the IBridgeReceiver interface, which
currently provides the method receiveEthFromBridge for reacting to WETH
(WRBTC) cross-back transfers. This can be extended in the future to also
account for token transfers.

The ERC777Recipient interface was not reused for this purpose, because it
doesn't define a payable method for reacting to transfers, and because
the method it provides includes parameters that we probably won't need.